Improved natla's pistols cabin dynamic object.
This is the only part of the game so far that room switching doesn't load the proper geometry. Apparently after putting all fuses the room should be replaced with a new one with the cabin as room surfaces but for some reason it doesn't load up. Saving and loading will load the correct room. For now the solution is to have the dynamic object with the entire cabin geometry which fixes the problem.